Preliminary version of the Aesmix tool
Aesmix is a library, written in C, that implements Mix&Slice encryption mode.
The implementation of Aesmix leverages AES as its building block.

To offer a more widespread access of the Mix&Slice capabilities and since
MOSAICrOWN policy engine has been implemented in Python, we provide a Python
wrapper to Aesmix library.